translation of all files
