+++ SudokuSolver.py
@@ -3,11 +3,11 @@ import sys
 import time
 import random
 from BruteForce import BruteForce
-from RecursiveSolver import BackTracking
+from BackTracking import BackTracking
 
 class SudokuSolver:
     def __init__(self):
-        # Définition des couleurs
+        # Defining colors
         self.WHITE = (255, 255, 255)
         self.BLACK = (0, 0, 0)
         self.GRAY = (200, 200, 200)
@@ -15,22 +15,22 @@ class SudokuSolver:
         self.RED = (255, 0, 0)
         self.input_value = None
 
-        # Initialisation de Pygame
+        # Initializing Pygame
         pygame.init()
 
-        # Définition de la taille de l'écran
+        # Setting screen size
         self.WIDTH, self.HEIGHT = 540, 600
         self.SCREEN = pygame.display.set_mode((self.WIDTH, self.HEIGHT))
         pygame.display.set_caption("Sudoku Solver")
 
-        # Initialisation des variables
+        # Initializing variables
         self.sudoku_grid = None
         self.selected_button = None
         self.solving_method = None
         self.solve_time = None
 
     def load_sudoku(self, filename):
-        # Charger une grid Sudoku à partir d'un fichier texte
+        # Load a Sudoku grid from a text file
         sudoku_grid = []
         with open(filename, 'r') as file:
             for line in file:
@@ -39,7 +39,7 @@ class SudokuSolver:
         return sudoku_grid
 
     def draw_sudoku(self, grid, missing_numbers=None):
-        # Dessiner la grid Sudoku sur l'écran
+        # Draw the Sudoku grid on the screen
         cell_size = self.WIDTH // 9
         for i in range(9):
             for j in range(9):
@@ -47,17 +47,17 @@ class SudokuSolver:
                 color = self.BLACK
                 if missing_numbers and (i, j) in missing_numbers:
                     color = self.RED
-                # Dessiner la valeur du chiffre s'il est différent de zéro
+                # Draw the digit value if it's not zero
                 if cell_value != 0:
                     font = pygame.font.Font(None, 36)
                     text_surface = font.render(str(cell_value), True, color)
                     text_rect = text_surface.get_rect(center=(j * cell_size + cell_size // 2, i * cell_size + cell_size // 2))
                     self.SCREEN.blit(text_surface, text_rect)
-                # Dessiner la grid
+                # Draw the grid
                 pygame.draw.rect(self.SCREEN, self.BLACK, (j * cell_size, i * cell_size, cell_size, cell_size), 1)
 
     def draw_buttons(self):
-        # Dessiner les boutons pour sélectionner les grids Sudoku
+        # Draw buttons to select Sudoku grids
         font = pygame.font.Font(None, 36)
         button_labels = ["Sudoku 1", "Sudoku 2", "Sudoku 3", "Sudoku 4", "Sudoku 5", "Brute Force"]
         button_rects = []
@@ -74,7 +74,7 @@ class SudokuSolver:
         return button_rects
 
     def solve_with_brute_force(self, grid, empty):
-        # Résoudre la grid Sudoku en utilisant la méthode de la force brute
+        # Solve the Sudoku grid using the brute force method
         start_time = time.time()
         solver = BruteForce(grid)
         solver.add_empty(empty)
@@ -85,7 +85,7 @@ class SudokuSolver:
             return 0, False
 
     def solve_with_recursion(self, grid):
-        # Résoudre la grid Sudoku en utilisant la méthode de la récursion
+        # Solve the Sudoku grid using the recursion method
         start_time = time.time()
         solver = BackTracking(grid)
         if solver.solve_sudoku():
@@ -103,8 +103,8 @@ class SudokuSolver:
         return missing_numbers
 
     def run(self):
-        # Lancer le programme principal
-        start_program_time = time.time()  # Enregistrer le temps de départ du programme
+        # Run the main program
+        start_program_time = time.time()  # Record program start time
         running = True
         missing_numbers = None
         while running:
@@ -114,11 +114,11 @@ class SudokuSolver:
                     running = False
                 elif event.type == pygame.KEYDOWN:
                     if pygame.K_1 <= event.key <= pygame.K_9:
-                        self.input_value = event.key - pygame.K_0 # Convertir la touche en chiffre
-                        print (self.input_value)
+                        self.input_value = event.key - pygame.K_0 # Convert key to digit
+                        print(self.input_value)
                     elif event.key == pygame.K_RETURN and selected_file == "resolved_matrix1.txt":
                         self.solving_method = self.solve_with_brute_force
-                        # Traiter l'entrée ici si nécessaire
+                        # Process input here if needed
                 elif event.type == pygame.MOUSEBUTTONDOWN:
                     if event.button == 1:
                         x,y = event.pos
@@ -153,7 +153,7 @@ class SudokuSolver:
                                 self.solving_method = None
                                 self.solve_time = None
                                 missing_numbers = None
-                        if event.button == 1:  # L'utilisateur clique avec le bouton gauche de la souris
+                        if event.button == 1:
                             cell_size = self.WIDTH // 9
 
             if self.sudoku_grid:
@@ -169,22 +169,21 @@ class SudokuSolver:
                     if success:
                         self.draw_sudoku(self.sudoku_grid, missing_numbers)
                     else:
-                        print("La résolution a échoué.")
+                        print("Resolution failed.")
                 if missing_numbers:
                     self.draw_sudoku(self.sudoku_grid, missing_numbers)
                 else:
                     self.draw_sudoku(self.sudoku_grid)
                 font = pygame.font.Font(None, 36)
-                font_temps = pygame.font.Font(None, 18)
+                font_time = pygame.font.Font(None, 18)
                 text_surface1 = font.render("Recursion", True, self.BLACK)
                 text_surface2 = font.render("BruteForce", True, self.BLACK)
-                text_surface3 = font.render("Retour", True, self.BLACK)
+                text_surface3 = font.render("Back", True, self.BLACK)
                 text_rect1 = text_surface1.get_rect(center=(self.WIDTH // 4 - 50, self.HEIGHT - 30))
                 text_rect2 = text_surface2.get_rect(center=(self.WIDTH // 4 - 50, self.HEIGHT - 30))
                 text_rect3 = text_surface3.get_rect(center=(self.WIDTH // 1.9, self.HEIGHT - 30))
                 pygame.draw.rect(self.SCREEN, self.RED, (text_rect3.x - 10, text_rect3.y - 10, text_rect3.width + 20, text_rect3.height + 20))
                 if selected_file == "resolved_matrix1.txt":
-                    """self.SCREEN.blit(text_surface2, text_rect2)"""
                     if self.input_value is not None:
                         text_surface = pygame.font.Font(None, 36).render(str(self.input_value), True, self.BLACK)
                         text_rect = text_surface.get_rect(center=(self.WIDTH // 4 - 50, self.HEIGHT - 30))
@@ -195,8 +194,8 @@ class SudokuSolver:
                 self.SCREEN.blit(text_surface3, text_rect3)
             
                 if self.solve_time is not None:
-                    solve_time_text = f"Temps de résolution: {self.solve_time:.4f} s"
-                    solve_time_surface = font_temps.render(solve_time_text, True, self.BLACK)
+                    solve_time_text = f"Solving Time: {self.solve_time:.4f} s"
+                    solve_time_surface = font_time.render(solve_time_text, True, self.BLACK)
                     solve_time_rect = solve_time_surface.get_rect(center=(self.WIDTH // 1.9 + 160, self.HEIGHT - 28))
                     self.SCREEN.blit(solve_time_surface, solve_time_rect)
