+++ backtracking.py
@@ -1,68 +1,77 @@
-from ConvertRaw import  ConvertRaw
+import time
 
-class BackTracking :
+class BackTracking:
+    def __init__(self, grid):
+        self.grid = grid
 
-    def __init__ (self, file_path) :
-        self.mat = [
-            ["_", "7", "2", "9", "_", "_", "_", "3", "_"], 
-            ["_", "_", "1", "_", "_", "6", "_", "8", "_"], 
-            ["_", "_", "_", "_", "4", "_", "_", "6", "_"], 
-            ["9", "6", "_", "_", "_", "4", "1", "_", "8"], 
-            ["_", "4", "8", "7", "_", "5", "_", "9", "6"], 
-            ["_", "_", "5", "6", "_", "8", "_", "_", "3"], 
-            ["_", "_", "_", "4", "_", "2", "_", "1", "_"], 
-            ["8", "5", "_", "_", "6", "_", "3", "2", "7"], 
-            ["1", "_", "_", "8", "5", "_", "_", "_", "_"]
-            ]
+    #display the sudoku grid
+    def display_grid(self):
+        for i in range(9):
+            if i % 3 == 0 and i != 0:
+                print("- - - - - - - - - - - -")
 
-    #display the game with an easely readable format
-    def display_terminal (self):
-        test = 0
-        print ("----------------------")
-        for colum in self.jeu_01 :
-            for line in colum :
-                if test < 8 :
-                    if test == 2 or test == 5:
-                        print (str (line) + " | " ,end='')
-                    else :
-                        print (str (line) + " " ,end='')
-                    test += 1
-                else :
-                    print (line)
-                    print ("----------------------")
-                    test = 0
+            for j in range(9):
+                if j % 3 == 0 and j != 0:
+                    print(" | ", end="")
 
-    #try to find an empty number and return its coordinate or return False if thes is none
-    def find_empty (self) :
-        for row in range (len (self.mat)) :
-            for col in range (len (self.mat [row])) :
-                if self.mat [row] [col] == "_" :
-                    return (row, col)
-        
-        return False
+                if j == 8:
+                    print(self.grid[i][j])
+                else:
+                    print(str(self.grid[i][j]) + " ", end="")
 
-    #verify the validity of the given number using its coordinate
-    def is_valid(self, number, row, col):
-        number = str (number)
-        for nimber_row in self.mat[row] :
-            if number == nimber_row[0] :
+    #find empty spots to fill them
+    def find_empty_spotes(self):
+        for i in range(9):
+            for j in range(9):
+                if self.grid[i][j] == 0:
+                    return i, j
+        return None
+
+    #test if the gicven number is valid
+    def valid_number(self, ligne, colonne, chiffre):
+        for i in range(9):
+            if self.grid[ligne][i] == chiffre or self.grid[i][colonne] == chiffre or self.grid[(ligne // 3) * 3 + i // 3][(colonne // 3) * 3 + i % 3] == chiffre:
                 return False
-            
-        for nimber_col in self.mat :
-            if number == nimber_col[col][0] :
-                return False
-            
-        for i in range(3):
-            for j in range(3):
-                if(self.mat[i + row][j + col] == number):
-                    return True
-        return False
-        
         return True
 
-    #solving the sudoku
-    def solve (self) :
-        
+    #solve the sudoku game
+    def solve_sudoku(self):
+        case_vide = self.find_empty_spotes()
+        if not case_vide:
+            return True
+
+        ligne, colonne = case_vide
+
+        for chiffre in range(1, 10):
+            if self.valid_number(ligne, colonne, chiffre):
+                self.grid[ligne][colonne] = chiffre
+
+                if self.solve_sudoku():
+                    return True
+
+                self.grid[ligne][colonne] = 0
+
+        return False
+
 
 if __name__ == "__main__":
-    sudoku = BackTracking()
+    fichier_sudoku = input("Entrez le nom du fichier Sudoku : ")
+
+    try:
+        with open(fichier_sudoku, "r") as file:
+            my_matrix = [[int(num) if num != '_' else 0 for num in line.strip()] for line in file]
+
+        solveur = BackTracking(my_matrix)
+        print("grid de Sudoku à résoudre :")
+        solveur.display_grid()
+        print("\nRésolution en cours...\n")
+
+        start_time = time.time()
+        if solveur.solve_sudoku():
+            print("Sudoku Résolu :")
+            solveur.display_grid()
+            print("Temps d'exécution:", time.time() - start_time, "secondes")
+        else:
+            print("Pas de solution possible.")
+    except FileNotFoundError:
+        print("Le fichier spécifié n'existe pas.")
